Bug 1088: Fix CONFIG_ECMASCRIPT_SMJS_HEARTBEAT
Makefile.config.in ignored the value of
CONFIG_ECMASCRIPT_SMJS_HEARTBEAT chosen by configure and instead used
the value of CONFIG_ECMASCRIPT_SMJS for it too.  This broke building
on SpiderMonkey versions where JS_TriggerOperationCallback is not
available.

No NEWS entry because this bug was not in any ELinks release.