configure: Restore "x" in "[$]WITHVAL_$1" comparison
Fix a minor bug introduced in commit
de45a1df516deee13fe002a0f71678a2ce46cd73 on 2005-07-21.
Before that, EL_CHECK_OPTIONAL_LIBRARY used to do this:

	if test -n "$withval" && test "x$withval" != xno; then
		AC_MSG_ERROR([$2 not found])
	fi

It was then changed to look in "[$]WITHVAL_$1" rather than "$withval",
but the initial "x" was lost in that change.  Restore it.