[action] Added reopen-last-closed-tab action. Refs #309
The uri of the latest closed tab uri is remembered in struct terminal.
And when the reopen-last-closed-tab action is executed, then uri
is open in new tab in foreground. Only one is in memory.
If you close two tabs, only on one of them can be reopened.

No key is assigned to this action by default.