[integrity] store integrity value in url . Refs #284
For ecmascript it is '^integrity\0url'.
Weak points:
Bad script is not executed, but is stored in cache, so next requests will take it from cache, not from network.
Checksum is calculated everytime script is loaded (it can be many times per second).
Scripting pre_format_hook can modify the body of script and cause a wrong checksum.
Error message is written to stderr.