require  quick JS config file
you can merge it when the config file is merged upstream or build from
my "shared" branch instead quickjs@shared in my account .  if it is declined I will submit it to quickjs-ng instead.
complete madness to look in the dark for a library instead of making a
config file. double madness to use static libs also considering the size
of these libs. we try to reduce the packet size by keeping all files
separate instead of in a gigantic blob of code

cannot understand why this particular lib is made this way when the same
person has made ffmpeg that is famous for absolute zero tolerance for
static libs. talk about one eighty turn. it is impossible to build
ffmpeg if pkg config return a static lib

 # elinks ELF including JS down to one meg! (two with debugging )

when netsurf and quick JS is built shared the program file is only one
meg ! without debugging . compared to around four + two more from JS and netsurf libs. or twelve with debugging

this is for netsurf to also output shared files that are automatically
joined with the config file

 export COMPONENT_TYPE=lib-shared